move to binarySearch in den
